refactor: KVStorage
