+++ packages/shared-base/package.json
@@ -26,6 +26,7 @@
     "async-call-rpc": "^6.3.1",
     "bignumber.js": "9.1.1",
     "dompurify": "^3.0.5",
+    "history": "^5.3.0",
     "i18next-browser-languagedetector": "^7.1.0",
     "idb": "^7.1.1",
     "json-stable-stringify": "^1.0.2",
