+++ packages/shared-base/src/index.ts
@@ -16,10 +16,10 @@ export * from './Site/types.js'
 export * from './Sniffings/index.js'
 export * from './Subscriptions/index.js'
 export * from './i18n/index.js'
-export * from './kv-storage/index.js'
 export * from './onDemandWorker/index.js'
 export * from './SingletonModal/index.js'
 export * from './serializer/index.js'
 export * from './LegacySettings/index.js'
-export * from './settings/index.js'
+export * from './KVStorage/index.js'
+export * from './KVStorage/kv-storage/index.js'
 export * from './Telemetry/index.js'
